Refactoring for gosec future code
I moved the getAnnotationLevel into a new file called annotations_levels.js because I will use this function when generating the gosec report too.

I made the test/annotation.levels.test.js file up to date with the new changes.

And in the end I renamed the test/report.test.js because in future I will create a gosec.report.test.js file and I think the name test/bandit.report.test.js shows clearer what are the tests for.
