+++ test/bandit.report.test.js
